added url atribution json file, some linting.

+++ bing_image_downloader/downloader.py
@@ -1,5 +1,9 @@
+"""
+@xvdp added filter args
+mod default output_dir
+"""
 from typing import Optional
-import sys
+import os
 import shutil
 from pathlib import Path
 
@@ -11,7 +15,7 @@ except ImportError:  # Python 3
 
 def download(query: str,
              limit: int = 100,
-             output_dir: str = 'dataset',
+             output_dir: str = '.',
              adult_filter_off: bool = True,
              force_replace: bool = False,
              timeout: int = 60,
@@ -20,9 +24,8 @@ def download(query: str,
              size: Optional[str] = None,
              aspect: Optional[str] = None,
              people: Optional[str] = None,
-             verbose=True) -> None:
+             verbose: bool = True) -> None:
     """
-    added filter args
     filter args:
         color       None(all) or color, bw, RED, ORANGE, GREEN, YELLOW, TEAL, BLUE,
                                     PURPLE, BROWN, BLACK, GRAY, WHITE
@@ -34,31 +37,19 @@ def download(query: str,
 
         """
     # engine = 'bing'
-    if adult_filter_off:
-        adult = 'off'
-    else:
-        adult = 'on'
+    adult = 'off' if adult_filter_off else 'on'
 
     image_dir = Path(output_dir).joinpath(query).absolute()
 
-    if force_replace:
-        if Path.is_dir(image_dir):
-            shutil.rmtree(image_dir)
+    if force_replace and Path.is_dir(image_dir):
+        shutil.rmtree(image_dir)
+    os.makedirs(image_dir, exist_ok=True)
 
-    # check directory and create if necessary
-    try:
-        if not Path.is_dir(image_dir):
-            Path.mkdir(image_dir, parents=True)
-
-    except Exception as e:
-        print('[Error]Failed to create directory.', e)
-        sys.exit(1)
-
-    print("[%] Downloading Images to {}".format(str(image_dir.absolute())))
+    print(f"[%] Downloading Images to {image_dir.absolute()}")
     bing = Bing(query, limit, image_dir, adult, timeout, img_type, color,
                 size, aspect, people, verbose)
     bing.run()
 
 
 if __name__ == '__main__':
-    download('dog', output_dir="..\\Users\\cat", limit=10, timeout=1)
+    download('dog', output_dir="cat", limit=10, timeout=1)
